.azure-pipelines/repo_mgmt/requirements.txt
GitPython==3.1.40